Add files via upload
Pictures of finished game. 
"Chess Welcome Screen" is the screen the game opens up to. It allows user to choose who is black/white, the names of the players, and the desired duration of their game.
There are restrictions of a 10 character limit and only English letters on names.

"Chess In-Game" is the screen during the game. Possible moves are available when a piece is selected. The current player can restart the game, resign, or exit to main menu at any time. Below those options is each side's material count. A timer records how much time each side has remaining. There is a separate time for each player.

"Chess End Game" is the screen that displays after the game ends. It displays who won, by what method, and allows the players to restart their game, return to the board, or exit to main menu to change settings.